fix: missing `isClassic` argument in `tx.toData()`
The missing `isClassic` flag leads to the tx method `toData()` interpreting the tx as message as if it were a phoenix-1 msg. This panics - because on phoenix-1 oracle messages are not supported. This error inhibits the actual broadcast error message (which would be otherwise pretty useful to have in the logs).